+++ NEWS.md
@@ -1,3 +1,8 @@
+Version 0.7.2
+================================================================================
+
+- Add includewithdrawn flag to nhanesSearchTableNames (thanks @Ifitrains https://github.com/cjendres1/nhanes/issues/15). Updated README. 
+
 Version 0.7.1
 ================================================================================